Move tests to feature folder
