+++ tests/Feature/UserTest.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Tests;
+namespace Tests\Feature;
 
 use App\User;
+use Tests\IntegrationTestCase;
 
 class UserTest extends IntegrationTestCase
 {
